hotel-api/requirements.txt
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
Flask-RESTful==0.3.9
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
PyMySQL==1.0.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2